+++ projetobackend/src/main/java/com/senac/daht/dto/response/PersonagemDTOResponse.java
@@ -1,11 +1,13 @@
 package com.senac.daht.dto.response;
 
 
+import jakarta.validation.constraints.NotNull;
+
 public class PersonagemDTOResponse {
     private Long id;
-    private int vida;
-    private int ouro;
-    private int xp;
+    private Double vida;
+    private Double ouro;
+    private Double xp;
     private String nomeUsuario;
 
     public Long getId() {
@@ -16,27 +18,27 @@ public class PersonagemDTOResponse {
         this.id = id;
     }
 
-    public int getVida() {
+    public Double getVida() {
         return vida;
     }
 
-    public void setVida(int vida) {
+    public void setVida(Double vida) {
         this.vida = vida;
     }
 
-    public int getOuro() {
+    public Double getOuro() {
         return ouro;
     }
 
-    public void setOuro(int ouro) {
+    public void setOuro(Double ouro) {
         this.ouro = ouro;
     }
 
-    public int getXp() {
+    public Double getXp() {
         return xp;
     }
 
-    public void setXp(int xp) {
+    public void setXp(Double xp) {
         this.xp = xp;
     }
 
@@ -47,4 +49,5 @@ public class PersonagemDTOResponse {
     public void setNomeUsuario(String nomeUsuario) {
         this.nomeUsuario = nomeUsuario;
     }
+
 }
