+++ projetobackend/src/main/java/com/senac/daht/dto/response/RegistroOuroDTOResponse.java
@@ -1,21 +1,21 @@
 package com.senac.daht.dto.response;
 
 public class RegistroOuroDTOResponse {
-    private int id;
-    private int quantidade;
+    private Long id;
+    private Double quantidade;
     private int idPersonagem;
 
     // Getters e Setters
-    public int getId() {
+    public Long getId() {
         return id;
     }
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
-    public int getQuantidade() {
+    public Double getQuantidade() {
         return quantidade;
     }
-    public void setQuantidade(int quantidade) {
+    public void setQuantidade(Double quantidade) {
         this.quantidade = quantidade;
     }
     public int getIdPersonagem() {
